requirements.txt
fastapi==0.123.10
uvicorn==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1
google-generativeai==0.8.3
aiofiles==23.2.1
